Squish a level of error types in the parser
We have to deal with the fact that
Parsec's errors are `ParseError`s while ours are `String`s,
so they don't belong to the same `Monad Either α`.
Previously, we exposed that publicly.
Now we don't.

The downside is you can't get the raw error if you actually want it.
The advantage is that it's easier to compose things monadically,
and the errors probably wouldn't have been useful anyway.